Split out getServerContextValue (#108)
